requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
pandas
scikit-learn
joblib
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
passlib[bcrypt]
python-multipart
python-jose[cryptography]
email-validator
bcrypt==4.0.1